refactor(stella-pipeline): split the #1701 ladder tests into verify/tests/uncollected.rs
`verify/tests.rs` was 1415 lines — one edit from the 1500-line ratchet, which
takes no new baseline entries. The rung's cases get their own module rather
than pushing the parent over, and they share a premise the rest of the ladder
tests do not: every probe in them worked.

Also folds the doc list's `3b.` back into item 3 (rustdoc reads a `3b.`
continuation as a lazy list item, which `-D clippy::doc_lazy_continuation`
rejects) — the two routes to the abstain rung are one state, so one entry.
